Add LRU memory cache to speed up access

+++ cache/cache.go
@@ -6,6 +6,7 @@ import (
 	"time"
 
 	"gitgud.io/softashell/comfy-translator/translator"
+	lru "github.com/hashicorp/golang-lru"
 	_ "github.com/mattn/go-sqlite3" // Sql driver
 	log "github.com/sirupsen/logrus"
 )
@@ -23,23 +24,26 @@ const (
 )
 
 type Cache struct {
-	db *sql.DB
+	db       *sql.DB
+	lrustore map[string]*lru.TwoQueueCache
 }
 
 type Item struct {
-	Text        string `storm:"id"` // primary key
 	Translation string
-	ErrorCode   translationError `storm:"index"`
+	ErrorCode   translationError
 	ErrorText   string
 	Timestamp   int64
 }
 
-func NewCache(filePath string, cacheSize int) (*Cache, error) {
+func NewCache(filePath string, cacheSize int, translators []string) (*Cache, error) {
 	db, err := sql.Open("sqlite3", filePath+"?cache=shared&mode=rwc&_synchronous=1&_auto_vacuum=2&_journal_mode=WAL")
 	if err != nil {
 		log.Fatal(err)
 		return nil, err
 	}
+
+	// Avoids database is locked errors
+	db.SetMaxOpenConns(1)
 
 	if _, err := db.Exec(fmt.Sprintf("PRAGMA cache_size = -%d", cacheSize)); err != nil {
 		log.Errorf("Failed to increase page cache size! %s", err)
@@ -47,7 +51,17 @@ func NewCache(filePath string, cacheSize int) (*Cache, error) {
 		log.Infof("Page cache set to %d kib", cacheSize)
 	}
 
-	cache := &Cache{db: db}
+	store := make(map[string]*lru.TwoQueueCache)
+	for _, t := range translators {
+		s, err := lru.New2Q(5000)
+		if err != nil {
+			log.Fatalf("Can't start memory cache for %s", t)
+		}
+
+		store[t] = s
+	}
+
+	cache := &Cache{db: db, lrustore: store}
 
 	cache.migrateDatabase()
 
@@ -87,26 +101,46 @@ func (c *Cache) Put(bucketName, text, translation string, cerr error) error {
 		log.Fatal(err)
 	}
 
+	// Add to memory cache
+	i := Item{
+		Translation: translation,
+		ErrorCode:   errorCode,
+		ErrorText:   errorText,
+		Timestamp:   time.Now().UTC().Unix(),
+	}
+
+	c.lrustore[bucketName].Add(text, i)
+
 	return err
 }
 
 func (c *Cache) Get(bucketName, text string) (string, bool, error) {
-	stmt, err := c.db.Prepare("SELECT id, translation, errorCode, errorText, time FROM Translations WHERE service = ? AND text = ?")
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer stmt.Close()
-
 	var found bool
 	var id int64
 	var translation string
 	var errorCode translationError
 	var errorText string
 	var timestamp int64
+	var err error
 
-	err = stmt.QueryRow(bucketName, text).Scan(&id, &translation, &errorCode, &errorText, &timestamp)
-	if err != nil {
-		return translation, found, nil
+	item, ok := c.lrustore[bucketName].Get(text)
+	if ok {
+		i := item.(Item)
+
+		translation = i.Translation
+		errorCode = i.ErrorCode
+		errorText = i.ErrorText
+	} else {
+		stmt, err := c.db.Prepare("SELECT id, translation, errorCode, errorText, time FROM Translations WHERE service = ? AND text = ?")
+		if err != nil {
+			log.Fatal(err)
+		}
+		defer stmt.Close()
+
+		err = stmt.QueryRow(bucketName, text).Scan(&id, &translation, &errorCode, &errorText, &timestamp)
+		if err != nil {
+			return translation, found, nil
+		}
 	}
 
 	if translation != "" {
@@ -117,9 +151,11 @@ func (c *Cache) Get(bucketName, text string) (string, bool, error) {
 		errorTime := time.Unix(timestamp, 0)
 
 		if time.Since(errorTime) > getCacheExpiration(errorCode) {
-			_, err = c.db.Exec(fmt.Sprintf("DELETE FROM Translations WHERE id = %d", id))
-			if err != nil {
-				log.Warn("unable to delete item: ", err)
+			if id != 0 {
+				_, err = c.db.Exec(fmt.Sprintf("DELETE FROM Translations WHERE id = %d", id))
+				if err != nil {
+					log.Warn("unable to delete item: ", err)
+				}
 			}
 
 			// Act as if nothing was found
